Further break apart DeepLink.swift
